requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
langchain>=0.1.20
langchain-community>=0.0.38
langchain-groq>=0.1.3
chromadb==0.4.18
sentence-transformers>=2.3.0
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4
sqlalchemy==2.0.23

